# PERF: Index.take to check for full range indices