--- a/pandas/core/indexes/multi.py
+++ b/pandas/core/indexes/multi.py
@@ -2247,7 +2247,7 @@ class MultiIndex(Index):
         # only fill if we are passing a non-None fill_value
         allow_fill = self._maybe_disallow_fill(allow_fill, fill_value, indices)
 
-        if lib.is_range_indexer(indices, len(self)):
+        if indices.ndim == 1 and lib.is_range_indexer(indices, len(self)):
             return self.copy()
 
         na_value = -1